Rename object.utils -> mask-object
