+++ shared/modules/mask-object.ts
